change the file.

+++ controleur/vers_la_validation_comment_table.php
@@ -1,6 +1,12 @@
 <?php
+require '/../functions.php';
 function inserer_comment_table() 
 { 
+	$errors = array();
+	$idBlog = htmlspecialchars($_GET['blog_post_id']);
+	$auteurComment = htmlspecialchars($_GET['auteur']);
+	$message = htmlspecialchars($_GET['message']);
+
 	function chargerClasse($Comment)
 	{
   		require  'modele/' .$Comment.'.php';
@@ -9,30 +15,57 @@ function inserer_comment_table()
 
 	$bdd = new PDO('mysql:host=localhost;dbname=blo;charset=utf8', 'root', '');
 
-	$_GET['blog_post_id'] = (int) $_GET['blog_post_id'];
-	$blogPostId = $_GET['blog_post_id'];
+	$idBlog = (int) $idBlog;
 
-	$_GET['id'] = (int) $_GET['id'];
-	$id = $_GET['id'];
-	
-	$Comment = new Comment();
-	$Comment->setBlog_post_id($blogPostId);
-    $Comment->setAuteur($_GET['auteur']);  
-    $Comment->setMessage($_GET['message']);  
+	$BlogPostManager = new BlogPostManager($bdd);
+			
+	$blog = $BlogPostManager->read($idBlog);
+	$idDataBase = $blog->id();
 
-	$CommentManager = new CommentManager($bdd);
-
-	$birth = $CommentManager->birth($Comment, $blogPostId);
-
-	if ($birth == true)
+	/*$_GET['id'] = (int) $_GET['id'];
+	$id = $_GET['id'];*/
+	if (isset($idBlog) AND !empty($idBlog) AND $idBlog == $idDataBase)
 	{
-		$delete = $CommentManager->delete($id);
+		if (isset($auteurComment) AND !empty($auteurComment))
+		{
+			if (isset($message) AND !empty($message))
+			{
+				$Comment = new Comment();
+				$Comment->setBlog_post_id($idBlog);
+    			$Comment->setAuteur($auteurComment);  
+    			$Comment->setMessage($message);  
+
+				$CommentManager = new CommentManager($bdd);
+
+				$birth = $CommentManager->birth($Comment, $blogPostId);
+
+				if ($birth == true)
+				{
+					$delete = $CommentManager->delete($id);
+				}
+				else
+				{
+					include(dirname(__FILE__).'/../vue/admin/menu_partie_admin.php');
+				}
+
+				include(dirname(__FILE__).'/../vue/admin/comment_definitivement_valider.php');
+			}
+			else
+			{
+   				$errors['message'] = "Le commentaire n'est pas valide !!!";
+        		debug($errors);
+			}
+		}
+		else
+		{
+   			$errors['auteur'] = "L'auteur n'est pas valide !!!";
+        	debug($errors);
+		}
 	}
 	else
 	{
-		include(dirname(__FILE__).'/../vue/admin/menu_partie_admin.php');
+   		$errors['blog_post_id'] = "L'id n'existe pas donc le blog demandé n'existe pas !!!";
+        debug($errors);
 	}
-
-	include(dirname(__FILE__).'/../vue/admin/comment_definitivement_valider.php');
 } 
 ?>
